Expects "id" attributes to be numbers rather than valid ids.
(Valid ids are strings and have quotes in the json. Apparently when we do this hack where we just create an attribute that stores the id, rather than setting up a relationship, it gets sent over the API as a number rather than a string.)

This is getting fixed now because we just fixed the rails API validator so that it catches this issue.