Format id attributes in *request* as actual ids
Ugh, this is why we should be using proper relationships, not sending ids as attributes. Currently, this happens to work because automatic type conversion is kind to us. But the backend uses the attributes as numbers, and the frontend uses them as strings. Hence we have to format them differently in the request and response.

In any case, this gets the fixtures where they need to be. Cleaning this up properly will require refactoring the code to use proper relationships.